+++ src/dividat-driver/server/main.go
@@ -22,11 +22,11 @@ var version string
 const serverPort = "8382"
 
 // Start the driver server
-func Start(logger *logrus.Logger) context.CancelFunc {
+func Start(logger *logrus.Logger, origins []string) context.CancelFunc {
 	// Log Server
 	logServer := logging.NewLogServer()
 	logger.AddHook(logServer)
-	http.Handle("/log", logServer)
+	http.Handle("/log", corsHeaders(origins, logServer))
 
 	baseLog := logger.WithFields(logrus.Fields{
 		"version":        version,
@@ -51,17 +51,17 @@ func Start(logger *logrus.Logger) context.CancelFunc {
 
 	// Setup Senso
 	sensoHandle := senso.New(ctx, baseLog.WithField("package", "senso"))
-	http.Handle("/senso", sensoHandle)
+	http.Handle("/senso", corsHeaders(origins, sensoHandle))
 
 	// Setup SensingTex reader
 	flexHandle := flex.New(ctx, baseLog.WithField("package", "flex"))
-	http.Handle("/flex", flexHandle)
+	http.Handle("/flex", corsHeaders(origins, flexHandle))
 
 	// Setup RFID scanner
 	rfidHandle := rfid.NewHandle(ctx, baseLog.WithField("package", "rfid"))
 	// net/http performs a redirect from `/rfid` if only `/rfid/` is mounted
-	http.Handle("/rfid", rfidHandle)
-	http.Handle("/rfid/", rfidHandle)
+	http.Handle("/rfid", corsHeaders(origins, rfidHandle))
+	http.Handle("/rfid/", corsHeaders(origins, rfidHandle))
 
 	// Create a logger for server
 	log := baseLog.WithField("package", "server")
@@ -80,11 +80,10 @@ func Start(logger *logrus.Logger) context.CancelFunc {
 		"os":        systemInfo.Os,
 		"arch":      systemInfo.Arch,
 	})
-	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*")
+	http.Handle("/", corsHeaders(origins, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
 		w.Write(rootMsg)
-	})
+	})))
 
 	// Start the server
 	log.WithField("port", serverPort).Info("Starting HTTP server.")
@@ -106,4 +105,33 @@ func Start(logger *logrus.Logger) context.CancelFunc {
 	}()
 
 	return cancel
+}
+
+// Middleware for CORS headers, to be applied to any route that should be accessible from browser apps.
+func corsHeaders(origins []string, next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if len(r.Header["Origin"]) == 1 && contains(origins, r.Header["Origin"][0]) {
+			w.Header().Set("Access-Control-Allow-Origin", r.Header["Origin"][0])
+			w.Header().Set("Access-Control-Allow-Private-Network", "true")
+		}
+
+		// Announce that `Origin` header value may affect response
+		w.Header().Set("Vary", "Origin")
+
+		if r.Method == "OPTIONS" {
+			w.WriteHeader(200)
+			return
+		} else {
+			next.ServeHTTP(w, r)
+		}
+	})
+}
+
+func contains(slice []string, candidate string) bool {
+    for _, member := range slice {
+        if member == candidate {
+            return true
+        }
+    }
+    return false
 }