+++ Changelog.md
@@ -5,8 +5,13 @@
 ### Added
 
 - Add a new bridge for Senso Flex devices
+- Add CORS response headers for private network access
 - Tool to record data from Senso
 - Command-line interface to update firmware on Senso via Ethernet
+
+### Changed
+
+- Limit permissible CORS origins
 
 ### Removed